Embedded the resolverUtils logic to build and impact utilities
